requirements.txt
pytesseract 
opencv-python 
pillow
pyflowchart
pycfg